Global Variables added for wasy movement between pages (2) -shuvo

+++ back-end/medi-help-back-end/src/main/java/com/amakakeru/medihelpbackend/entity/Thread.java
@@ -1,11 +1,8 @@
 package com.amakakeru.medihelpbackend.entity;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.persistence.*;
 import lombok.Data;
-import org.springframework.format.annotation.DateTimeFormat;
 
-import java.sql.Timestamp;
 import java.util.Date;
 
 @Entity
@@ -15,7 +12,7 @@ public class Thread {
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
     @Column(name = "thread_id")
-    private Integer threadId;
+    private Long threadId;
 
     @Column(name = "uploader_id")
     private String uploaderId;
